igCauldron IGZ stuff

+++ igLibrary/Core/sizeofSize.cs
@@ -4,12 +4,12 @@ namespace igLibrary.Core
 	sealed class sizeofSize : Attribute
 	{
 		public uint _applicableVersion;
-		public uint _size32;
-		public uint _size64;
+		public ushort _size32;
+		public ushort _size64;
 		public IG_CORE_PLATFORM[] _platform;
 		
 		// This is a positional argument
-		public sizeofSize(uint applicableVersion, uint size32, uint size64, params IG_CORE_PLATFORM[] platform)
+		public sizeofSize(uint applicableVersion, ushort size32, ushort size64, params IG_CORE_PLATFORM[] platform)
 		{
 			this._size32 = size32;
 			this._size64 = size64;
